Feature: Extract ProviderMock to own file to pass CS check
