+++ tests/Test/Provider/AbstractProviderTest.php
@@ -7,68 +7,15 @@
 namespace Test\Provider;
 
 use SocialConnect\Common\Http\Client\ClientInterface;
-use SocialConnect\Provider\AbstractBaseProvider;
-use SocialConnect\Provider\AccessTokenInterface;
 use SocialConnect\Provider\Consumer;
 use SocialConnect\Provider\Session\SessionInterface;
 use Test\TestCase;
-
-class AbstractProviderMock extends AbstractBaseProvider {
-
-    /**
-     * @return string
-     */
-    public function getBaseUri()
-    {
-        // TODO: Implement getBaseUri() method.
-    }
-
-    /**
-     * Return Provider's name
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return 'fake';
-    }
-
-    /**
-     * @param array $requestParameters
-     * @return \SocialConnect\Provider\AccessTokenInterface
-     */
-    public function getAccessTokenByRequestParameters(array $requestParameters)
-    {
-        // TODO: Implement getAccessTokenByRequestParameters() method.
-    }
-
-    /**
-     * @return string
-     */
-    public function makeAuthUrl()
-    {
-        // TODO: Implement makeAuthUrl() method.
-    }
-
-    /**
-     * Get current user identity from social network by $accessToken
-     *
-     * @param AccessTokenInterface $accessToken
-     * @return \SocialConnect\Common\Entity\User
-     *
-     * @throws \SocialConnect\Provider\Exception\InvalidResponse
-     */
-    public function getIdentity(AccessTokenInterface $accessToken)
-    {
-        // TODO: Implement getIdentity() method.
-    }
-}
 
 class AbstractProviderTest extends TestCase
 {
     /**
      * @param ClientInterface|null $httpClient
-     * @return AbstractProviderMock
+     * @return ProviderMock
      */
     protected function getAbstractProviderMock(ClientInterface $httpClient = null, SessionInterface $session = null)
     {
@@ -86,7 +33,7 @@ class AbstractProviderTest extends TestCase
                 ->getMock();
         }
 
-        return new AbstractProviderMock(
+        return new ProviderMock(
             $httpClient,
             $session,
             new Consumer(
